requirements.txt
dvc==3.25.0
dvclive
pandas
pyyaml